requirements.txt
aiohttp==3.8.4
boto3==1.26.165
botocore==1.29.165
elevenlabs==0.2.18
numpy==1.24.3
openai==0.27.6
playsound==1.3.0
pocketsphinx==5.0.1
protobuf==3.12.4
PyAudio==0.2.13
python-dotenv==1.0.0
Requests==2.31.0
sounddevice==0.4.6
soundfile==0.12.1
tensorflow==2.12.0
torch==2.0.1+cu117
vosk==0.3.45
